project--vc-list-files: use Git's sparse-index
When dealing with exceptionally large Git repositories, the performance
of `project-find-file` can suffer dramatically as the list of files is
collected for completion. This adds insult to injury when you consider
cases where the developer has configured the repository to use a sparse
checkout where the vast majority of these files are not even present on
disk and are not valid candidates for completion.

Conditionally pass along the `--sparse` option to `git-ls-files` to
coerce `project--vc-list-files` only consider files present in the
sparse index (and presumably on disk). Verify this later by filtering
the results of `git-ls-files` with `file-exists-p`. This also serves to
filter out the directory entries still returned by `git-ls-files` when
run in this mode and to handle systems where the installed
`git-ls-files` doesn't yet understand `--sparse`.